+++ modulo2/semana5/projeto/src/controller/postController.ts
@@ -6,15 +6,36 @@ export const getPostById = async (
    res: Response
 ) => {
    try {
+      let message = "Success!"
 
       const { id } = req.params
 
-      const taskWithUserInfo = await businessGetTaskById(id)
+      const queryResult: any = await connection("labook_posts")
+         .select("*")
+         .where({ id })
 
-      res.status(200).send(taskWithUserInfo)
+      if (!queryResult[0]) {
+         res.statusCode = 404
+         message = "Post not found"
+         throw new Error(message)
+      }
+
+      const post: Post = {
+         id: queryResult[0].id,
+         photo: queryResult[0].photo,
+         description: queryResult[0].description,
+         type: queryResult[0].type,
+         createdAt: queryResult[0].created_at,
+         authorId: queryResult[0].author_id,
+      }
+
+      res.status(200).send({ message, post })
 
    } catch (error) {
-      res.status(400).send(error.message)
+      let message = error.sqlMessage || error.message
+      res.statusCode = 400
+
+      res.send({ message })
    }
 }
 
@@ -23,20 +44,31 @@ export const createPost = async (
    res: Response
 ) => {
    try {
+      let message = "Success!"
 
-      const { title, description, deadline, authorId } = req.body
+      const { photo, description, type } = req.body
 
-      await businessCreateTask(
-         title,
-         description,
-         deadline,
-         authorId
-      )
+      const token: string = req.headers.authorization as string
 
-      res.status(201).end()
+      const tokenData: AuthenticationData = getTokenData(token)
+
+      const id: string = generateId()
+
+      await connection("labook_posts")
+         .insert({
+            id,
+            photo,
+            description,
+            type,
+            author_id: tokenData.id
+         })
+
+      res.status(201).send({ message })
 
    } catch (error) {
-      res.statusMessage = error.message
-      res.status(500).end()
+      let message = error.sqlMessage || error.message
+      res.statusCode = 400
+
+      res.send({ message })
    }
 }
